requirements.txt
# Core dependencies
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
numpy>=1.24.0

# Database and vector search
chromadb>=0.4.22

# Gemini API
google-generativeai>=0.3.0

# Document processing
PyPDF2>=3.0.0
pdf2image>=1.16.0
python-pptx>=0.6.21
Pillow>=10.0.0

# Utilities
python-dateutil>=2.8.2
validators>=0.22.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
